requirements.txt
streamlit==1.31.0
pandas==2.1.4
psycopg2-binary==2.9.9
python-dotenv==1.0.0
streamlit-authenticator==0.3.2
msal==1.26.0
requests==2.31.0
matplotlib>=3.7.0
openpyxl>=3.1.0
